requirements.txt
python-dotenv
Flask
sqlalchemy